renamed test files for clarity
