+++ tests/measurements/test_codecarbon_energy_meter.py
